base64.h对应64位的修改

+++ IDCardFdv/base64.h
@@ -18,7 +18,7 @@ public:
 
 		out->resize(EncodedLength(in));
 
-		int input_len = in.size();
+		size_t input_len = in.size();
 		std::string::const_iterator input = in.begin();
 
 		while (input_len--) {
@@ -101,7 +101,7 @@ public:
 		unsigned char a3[3];
 		unsigned char a4[4];
 
-		int input_len = in.size();
+		size_t input_len = in.size();
 		std::string::const_iterator input = in.begin();
 
 		out->resize(DecodedLength(in));
@@ -196,7 +196,7 @@ public:
 		return (out == (out_begin + decoded_length));
 	}
 
-	static int DecodedLength(const char *in, size_t in_length) {
+	static size_t DecodedLength(const char *in, size_t in_length) {
 		int numEq = 0;
 
 		const char *in_end = in + in_length;
@@ -205,9 +205,9 @@ public:
 		return ((6 * in_length) / 8) - numEq;
 	}
 
-	static int DecodedLength(const std::string &in) {
+	static size_t DecodedLength(const std::string &in) {
 		int numEq = 0;
-		int n = in.size();
+		size_t n = in.size();
 
 		for (std::string::const_reverse_iterator it = in.rbegin(); *it == '='; ++it) {
 			++numEq;
@@ -216,11 +216,11 @@ public:
 		return ((6 * n) / 8) - numEq;
 	}
 
-	inline static int EncodedLength(size_t length) {
+	inline static size_t EncodedLength(size_t length) {
 		return (length + 2 - ((length + 2) % 3)) / 3 * 4;
 	}
 
-	inline static int EncodedLength(const std::string &in) {
+	inline static size_t EncodedLength(const std::string &in) {
 		return EncodedLength(in.length());
 	}
